Broke apart components of app.js into separate files. Also code cleanup.
